[54] RxForm Api improvements, adding custom effects
